Move Keycloak account client tests out of MFA
